+++ extension.js
@@ -1,6 +1,7 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 const vscode = require('vscode');
+const path = require('path');
 const fs = require('fs');
 let output = true;
 
@@ -26,7 +27,7 @@ function activate(context) {
 		// Display a message box to the user
 		console.log('Hello from hello!');
 
-		const filePath = uri.path.substring(1);
+		const Path = uri.path.substring(1);
 
 		/**
 		 * 将字符串的第一个字母大写
@@ -56,14 +57,45 @@ function activate(context) {
 			return roundedResult;
 		}
 
-		fs.stat(filePath, (err, stats) => {
+		fs.stat(Path, (err, stats) => {
 			if(err)
 			{
 				vscode.window.showInformationMessage(`Error${err}`);
 			}
 			if(stats.isDirectory())
 			{
-				vscode.window.showInformationMessage(`${capitalizeFirstLetter(filePath)} is not a file!`);
+				//vscode.window.showInformationMessage(`${capitalizeFirstLetter(Path)} is not a file!`);
+				function getFolderSize(folderPath) {
+					let totalSize = 0;
+					function getFileSize(filePath) {
+					const stat = fs.statSync(filePath);
+					if(stat.isFile())
+					{
+						totalSize += stat.size;
+					}
+					if(stat.isDirectory())
+					{
+						const files = fs.readdirSync(filePath);
+						files.forEach(file => {
+						  	const fullPath = path.join(filePath, file);
+						  	getFileSize(fullPath);
+						});
+					  }
+					}
+					getFileSize(folderPath);
+					return totalSize;
+				}
+				const size = getFolderSize(Path);
+				const creatTime = stats.birthtime.toLocaleString();
+				const modifyTime = stats.mtime.toLocaleString();
+				output = vscode.window.createOutputChannel('文件夹信息');
+				output.appendLine (`
+					文件夹大小为：${divideAndRound(size, 1024)}KiB;
+					文件夹创建于：${creatTime};
+					文件夹修改于：${modifyTime};
+					文件夹路径为：${capitalizeFirstLetter(Path)};`
+				);
+				output.show(true);
 			}
 			if(stats.isFile())
 			{
@@ -75,7 +107,7 @@ function activate(context) {
 					文件大小为：${divideAndRound(size, 1024)}KiB;
 					文件创建于：${creatTime};
 					文件修改于：${modifyTime};
-					文件路径为：${capitalizeFirstLetter(filePath)};`
+					文件路径为：${capitalizeFirstLetter(Path)};`
 				);
 				output.show(true);
 			}